Removed createHandle from the FlowStateMachine interface as it doesn't need to be publicly exposed
