+++ node/src/main/kotlin/net/corda/node/internal/CordaRPCOpsImpl.kt
@@ -1,5 +1,7 @@
 package net.corda.node.internal
 
+import com.google.common.util.concurrent.ListenableFuture
+import net.corda.client.rpc.notUsed
 import net.corda.core.contracts.Amount
 import net.corda.core.contracts.ContractState
 import net.corda.core.contracts.StateAndRef
@@ -13,14 +15,15 @@ import net.corda.core.node.NodeInfo
 import net.corda.core.node.services.NetworkMapCache
 import net.corda.core.node.services.StateMachineTransactionMapping
 import net.corda.core.node.services.Vault
+import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import net.corda.node.services.api.ServiceHubInternal
 import net.corda.node.services.messaging.requirePermission
 import net.corda.node.services.startFlowPermission
 import net.corda.node.services.statemachine.StateMachineManager
 import net.corda.node.utilities.transaction
-import org.bouncycastle.asn1.x500.X500Name
 import net.corda.nodeapi.CURRENT_RPC_USER
+import org.bouncycastle.asn1.x500.X500Name
 import org.jetbrains.exposed.sql.Database
 import rx.Observable
 import java.io.InputStream
@@ -100,14 +103,20 @@ class CordaRPCOpsImpl(
     override fun <T : Any> startTrackedFlowDynamic(logicType: Class<out FlowLogic<T>>, vararg args: Any?): FlowProgressHandle<T> {
         requirePermission(startFlowPermission(logicType))
         val currentUser = FlowInitiator.RPC(CURRENT_RPC_USER.get().username)
-        return services.invokeFlowAsync(logicType, currentUser, *args).createHandle(hasProgress = true) as FlowProgressHandle<T>
+        val stateMachine = services.invokeFlowAsync(logicType, currentUser, *args)
+        return FlowProgressHandleImpl(
+                id = stateMachine.id,
+                returnValue = stateMachine.resultFuture,
+                progress = stateMachine.logic.track()?.second ?: Observable.empty()
+        )
     }
 
     // TODO: Check that this flow is annotated as being intended for RPC invocation
     override fun <T : Any> startFlowDynamic(logicType: Class<out FlowLogic<T>>, vararg args: Any?): FlowHandle<T> {
         requirePermission(startFlowPermission(logicType))
         val currentUser = FlowInitiator.RPC(CURRENT_RPC_USER.get().username)
-        return services.invokeFlowAsync(logicType, currentUser, *args).createHandle(hasProgress = false)
+        val stateMachine = services.invokeFlowAsync(logicType, currentUser, *args)
+        return FlowHandleImpl(id = stateMachine.id, returnValue = stateMachine.resultFuture)
     }
 
     override fun attachmentExists(id: SecureHash): Boolean {
@@ -159,6 +168,33 @@ class CordaRPCOpsImpl(
                 is StateMachineManager.Change.Add -> StateMachineUpdate.Added(stateMachineInfoFromFlowLogic(change.logic))
                 is StateMachineManager.Change.Removed -> StateMachineUpdate.Removed(change.logic.runId, change.result)
             }
+        }
+    }
+
+    // I would prefer for [FlowProgressHandleImpl] to extend [FlowHandleImpl],
+    // but Kotlin doesn't allow this for data classes, not even to create
+    // another data class!
+    @CordaSerializable
+    private data class FlowHandleImpl<A>(
+            override val id: StateMachineRunId,
+            override val returnValue: ListenableFuture<A>) : FlowHandle<A> {
+
+        // Remember to add @Throws to FlowHandle.close if this throws an exception
+        override fun close() {
+            returnValue.cancel(false)
+        }
+    }
+
+    @CordaSerializable
+    private data class FlowProgressHandleImpl<A>(
+            override val id: StateMachineRunId,
+            override val returnValue: ListenableFuture<A>,
+            override val progress: Observable<String>) : FlowProgressHandle<A> {
+
+        // Remember to add @Throws to FlowProgressHandle.close if this throws an exception
+        override fun close() {
+            progress.notUsed()
+            returnValue.cancel(false)
         }
     }
 }
